Revert "fix: remove stale detours after inline labels"
This reverts commit fdee577dec27b04be383d62c406e9b08ab9d06a3.
