+++ tests/bug-reports/bug-report-20260901T055358Z/bug-report-20260901T055358Z.test.ts
@@ -8,16 +8,5 @@ test("bug-report-20260901T055358Z", () => {
 
   solver.solve()
 
-  const groundLabelConnector = solver
-    .netLabelToTraceSolver!.getOutput()
-    .traces.find(
-      (trace) => trace.mspPairId === "available-net-orientation-19-GND",
-    )
-  expect(groundLabelConnector).toBeDefined()
-  expect(groundLabelConnector!.tracePath).toHaveLength(2)
-  expect(groundLabelConnector!.tracePath.every((point) => point.y === 3)).toBe(
-    true,
-  )
-
   expect(solver).toMatchSolverSnapshot(import.meta.path)
 })
